Deprecate current RC list codebase
I'll be developing new one, which will use the common framework. I hope
it to be smaller.

This one will get deleted once new one is done. With this renaming I'll
be easy.
