+++ src/app/frontend/chrome/chrome.html
@@ -16,7 +16,7 @@ limitations under the License.
 
 <md-toolbar class="kd-toolbar">
   <div class="md-toolbar-tools kd-toolbar-tools">
-    <a ui-sref="replicationcontrollers" ui-sref-opts="{ reload: true }" tabindex="-1">
+    <a ui-sref="replicationcontrollersdeprecated" ui-sref-opts="{ reload: true }" tabindex="-1">
       <md-icon md-svg-icon="assets/images/kubernetes-logo.svg" class="kd-toolbar-logo"></md-icon>
     </a>
     <h2>
