+++ src/app/frontend/deploy/deployfromfile_controller.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {stateName as replicationcontrollerliststate} from 'replicationcontrollerlist/replicationcontrollerlist_state';
+import {stateName as replicationcontrollerliststate} from 'replicationcontrollerlistdeprecated/replicationcontrollerlist_state';
 
 /**
  * Controller for the deploy from file directive.
