+++ src/app/frontend/index_module.js
@@ -23,7 +23,7 @@ import indexConfig from './index_config';
 import routeConfig from './index_route';
 import logsModule from './logs/logs_module';
 import replicationControllerDetailModule from './replicationcontrollerdetail/replicationcontrollerdetail_module';
-import replicationControllerListModule from './replicationcontrollerlist/replicationcontrollerlist_module';
+import replicationControllerListModule from './replicationcontrollerlistdeprecated/replicationcontrollerlist_module';
 
 export default angular
     .module(
