+++ src/app/frontend/replicationcontrollerdetail/replicationcontrollerdetail.html
@@ -22,7 +22,7 @@ limitations under the License.
       <div class="kd-replicationcontrollerdetail-sidebar-header" layout-align="start center"
            layout="row">
         <div flex="none">
-          <md-button class="md-icon-button" ui-sref="replicationcontrollers">
+          <md-button class="md-icon-button" ui-sref="replicationcontrollersdeprecated">
             <md-icon md-font-library="material-icons">arrow_back</md-icon>
           </md-button>
         </div>
@@ -43,7 +43,7 @@ limitations under the License.
       <div class="kd-replicationcontrollerdetail-sidebar-header" layout-align="start center"
            layout="row">
         <div flex="none">
-          <md-button class="md-icon-button" ui-sref="replicationcontrollers">
+          <md-button class="md-icon-button" ui-sref="replicationcontrollersdeprecated">
             <md-icon md-font-library="material-icons">arrow_back</md-icon>
           </md-button>
         </div>
