+++ src/app/frontend/replicationcontrollerlistdeprecated/logsmenu.html
@@ -18,7 +18,7 @@ limitations under the License.
   <md-button ng-click="ctrl.openMenu($mdOpenMenu, $event)" class="kd-replicationcontroller-card-logs-button">
     Logs
   </md-button>
-  <md-menu-content class="kd-replicationcontroller-card-logs-menu" width="6">
+  <md-menu-content class="kd-replicationcontroller-card-logs-menu-deprecated" width="6">
     <md-menu-item class="kd-menu-logs-md-menu-item">
       <div>Logs</div>
     </md-menu-item>
