+++ src/app/frontend/replicationcontrollerlistdeprecated/logsmenu.scss
@@ -12,20 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import ReplicationControllerCardMenuController from './replicationcontrollercardmenu_controller';
+@import '../variables';
 
-/**
- * Returns directive definition object for logs menu.
- * @return {!angular.Directive}
- */
-export default function replicationControllerCardMenuDirective() {
-  return {
-    scope: {},
-    bindToController: {
-      'replicationController': '=',
-    },
-    controller: ReplicationControllerCardMenuController,
-    controllerAs: 'ctrl',
-    templateUrl: 'replicationcontrollerlist/replicationcontrollercardmenu.html',
-  };
+.kd-replicationcontroller-card-logs-menu-deprecated {
+  font-size: $body-font-size-base;
+}
+
+.kd-menu-logs-item-header {
+  box-sizing: inherit;
+  color: $foreground-2;
+  font-size: $body-font-size-base;
+  white-space: nowrap;
+}
+
+.kd-menu-logs-item {
+  box-sizing: inherit;
+  font-size: $body-font-size-base;
+  white-space: nowrap;
+}
+
+.kd-menu-logs-item-since {
+  width: 19 * $baseline-grid;
 }
