+++ src/app/frontend/replicationcontrollerlistdeprecated/logsmenu_controller.js
