+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollercard.scss
