+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollercard_controller.js
