+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollercardmenu.html
