+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollercardmenu.scss
