+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollercardmenu_controller.js
