+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollerlist_controller.js
