+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollerlistcontainer.html
@@ -13,12 +13,5 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 -->
-<md-button class="md-fab kd-replicationcontroller-deploy" ng-click="ctrl.redirectToDeployPage()">
-  <md-icon class="material-icons">add</md-icon>
-</md-button>
 
-<kd-replication-controller-list-container>
-  <kd-replication-controller-card ng-repeat="replicationController in ::ctrl.replicationControllers"
-                                  replication-controller="::replicationController">
-  </kd-replication-controller-card>
-</kd-replication-controller-list-container>
+<div class="kd-replication-controller-list-container-deprecated" layout-padding ng-transclude></div>
