+++ src/app/frontend/replicationcontrollerlistdeprecated/replicationcontrollerlistcontainer.scss
@@ -12,8 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/** Name of the state. Can be used in, e.g., $state.go method. */
-export const stateName = 'replicationcontrollers';
+@import '../variables';
 
-/** Absolute URL of the state. */
-export const stateUrl = '/replicationcontrollers';
+.kd-replication-controller-list-container-deprecated {
+  align-content: center;
+  align-items: center;
+  display: flex;
+  flex-flow: column wrap;
+}
