+++ src/app/frontend/replicationcontrollerlistdeprecated/zerostate/zerostate.html
@@ -39,7 +39,7 @@ limitations under the License.
         </div>
         <p class="kd-zerostate-lm kd-zerostate-rcs" ng-if="ctrl.containsOnlyKubeSystemRCs">
           Some system applications are running in the cluster.
-          <a class="md-raised md-primary" ui-sref="replicationcontrollers">Show them</a>
+          <a class="md-raised md-primary" ui-sref="replicationcontrollersdeprecated">Show them</a>
         </p>
       </md-card-content>
     </md-card>
