+++ src/app/frontend/replicationcontrollerlistdeprecated/zerostate/zerostate.scss
