+++ src/app/frontend/replicationcontrollerlistdeprecated/zerostate/zerostate_controller.js
