+++ src/app/frontend/replicationcontrollerlistdeprecated/zerostate/zerostate_state.js
