+++ src/test/frontend/common/filters/middleellipsis_filter_test.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import replicationcontrollerlistModule from 'replicationcontrollerlist/replicationcontrollerlist_module';
+import replicationcontrollerlistModule from 'replicationcontrollerlistdeprecated/replicationcontrollerlist_module';
 
 describe('Apply ellipsis filter', () => {
   const testedString = 'podName';
