+++ src/test/frontend/replicationcontrollerdetail/replicationcontrollerinfo_controller_test.js
@@ -14,7 +14,7 @@
 
 import ReplicationControllerInfoController from 'replicationcontrollerdetail/replicationcontrollerinfo_controller';
 import replicationControllerDetailModule from 'replicationcontrollerdetail/replicationcontrollerdetail_module';
-import {stateName as replicationcontrollers} from 'replicationcontrollerlist/replicationcontrollerlist_state';
+import {stateName as replicationcontrollers} from 'replicationcontrollerlistdeprecated/replicationcontrollerlist_state';
 
 describe('Replication Controller Detail controller', () => {
   /**
