+++ src/test/frontend/replicationcontrollerlistdeprecated/replicationcontrollerlistcontainer_test.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {binarySearchOptimalHeight, computeColumnCount} from 'replicationcontrollerlist/replicationcontrollerlistcontainer';
+import {binarySearchOptimalHeight, computeColumnCount} from 'replicationcontrollerlistdeprecated/replicationcontrollerlistcontainer';
 
 describe('Replication controller list container', () => {
